+++ src/lib/auth/Login.svelte
@@ -1,24 +1,25 @@
 <script lang="ts">
-	import { currentUser, pb } from './pocketbase';
-    import { useForm, validators, HintGroup, Hint, email, required } from 'svelte-use-form';
+	import { currentRole, currentUser, currentOrganization, pb } from '../pocketbase';
+	import { useForm, validators, HintGroup, Hint, email, required } from 'svelte-use-form';
 	import 'bootstrap-icons/font/bootstrap-icons.css';
 	const form = useForm();
 
-    var emailaddr =  "";
-    var password  = "";
+	var emailaddr = '';
+	var password = '';
 
-    async function login(){
-        await pb.collection('users').authWithPassword(emailaddr,password);
-    }
-    async function signout() {
-        pb.authStore.clear();
-    }
-
+	async function login() {
+		await pb.collection('users').authWithPassword(emailaddr, password);
+	}
+	async function signout() {
+		pb.authStore.clear();
+	}
 </script>
 
 {#if $currentUser}
-	<p>Signed in as {$currentUser.username}</p>
-    <button on:click={signout} > logout</button>
+	<p>
+		Signed in as {$currentUser.username}, {$currentRole},{$currentUser.org},{$currentOrganization}
+	</p>
+	<button on:click={signout}> logout</button>
 {:else}
 	<form on:submit|preventDefault>
 		<h1>Login</h1>
@@ -29,7 +30,7 @@
 			name="email"
 			id="email"
 			placeholder="goof@goofle.com"
-            bind:value={emailaddr}
+			bind:value={emailaddr}
 			use:validators={[required, email]}
 		/>
 		<HintGroup for="email">
@@ -43,7 +44,7 @@
 			name="password"
 			id="password"
 			placeholder="Password"
-            bind:value={password}
+			bind:value={password}
 			use:validators={[required]}
 		/>
 		<Hint for="password" on="required">This is a mandatory field</Hint>
